+++ libfsapfs/libfsapfs_container.c
@@ -623,7 +623,7 @@ int libfsapfs_container_open_file_io_handle(
 	static char *function                              = "libfsapfs_container_open_file_io_handle";
 	int bfio_access_flags                              = 0;
 	int file_io_handle_is_open                         = 0;
-	int file_io_handle_opened_in_library               = 0;
+	uint8_t file_io_handle_opened_in_library           = 0;
 
 	if( container == NULL )
 	{
